feat: use voyager transaction in transaction detail screen
chore: remove old implementation

chore: remove old list item

fix: network id

feat: show from / to for NFT transfers
