+++ packages/extension/src/background/transactions/transactionExecution.ts
@@ -36,7 +36,6 @@ export const executeTransaction = async (
   { wallet }: BackgroundService,
 ) => {
   const { transactions, abis, transactionsDetail } = action.payload
-  console.log(JSON.stringify(transactions, null, 2))
   if (!(await wallet.isSessionOpen())) {
     throw Error("you need an open session")
   }
@@ -62,7 +61,6 @@ export const executeTransaction = async (
     nonce,
     maxFee,
   })
-  console.log(JSON.stringify(transaction, null, 2))
 
   if (!checkTransactionHash(transaction.transaction_hash)) {
     throw Error("Transaction could not get added to the sequencer")
