+++ packages/extension/src/ui/features/accountActivity/TransactionDetail.tsx
@@ -1,19 +1,26 @@
 import { BigNumber } from "ethers"
+import { isString } from "lodash-es"
 import { FC, useMemo, useState } from "react"
-import styled from "styled-components"
+import styled, { useTheme } from "styled-components"
 
 import { IExplorerTransaction } from "../../../shared/explorer/type"
 import { Network } from "../../../shared/network"
 import { Token } from "../../../shared/token/type"
-import { entryPointToHumanReadable } from "../../../shared/transactions"
+import {
+  Transaction,
+  entryPointToHumanReadable,
+} from "../../../shared/transactions"
 import { CopyIconButton } from "../../components/CopyIconButton"
+import { CopyTooltip } from "../../components/CopyTooltip"
 import {
   Field,
   FieldGroup,
   FieldKey,
   FieldValue,
+  LeftPaddedField,
   SectionHeader,
 } from "../../components/Fields"
+import { ContentCopyIcon } from "../../components/Icons/MuiIcons"
 import { TransactionUnknownInline } from "../../components/Icons/TransactionUnknownInline"
 import { formatTruncatedAddress } from "../../services/addresses"
 import { formatDateTime } from "../../services/dates"
@@ -27,6 +34,7 @@ import { TokenField } from "../actions/transaction/fields/TokenField"
 import { TransactionDetailWrapper } from "./TransactionDetailWrapper"
 import {
   isNFTTransaction,
+  isNFTTransferTransaction,
   isSwapTransaction,
   isTokenMintTransaction,
   isTokenTransferTransaction,
@@ -37,6 +45,18 @@ import { NFTTitle } from "./ui/NFTTitle"
 import { TransactionCallDataBottomSheet } from "./ui/TransactionCallDataBottomSheet"
 import { TransactionIcon } from "./ui/TransactionIcon"
 import { TransferTitle } from "./ui/TransferTitle"
+
+function getErrorMessageFromErrorDump(errorDump?: string) {
+  try {
+    if (!isString(errorDump)) {
+      return undefined
+    }
+    const errorCode = errorDump.match(/^Error message: (.+)$/im)
+    return errorCode?.[1] ?? undefined
+  } catch {
+    return undefined
+  }
+}
 
 const StyledTransactionDetailWrapper = styled(TransactionDetailWrapper)`
   position: relative;
@@ -84,30 +104,66 @@ const HyperlinkText = styled.div`
   text-decoration: underline;
 `
 
-export interface IExplorerTransactionDetail {
-  explorerTransaction: IExplorerTransaction
-  explorerTransactionTransformed: TransformedTransaction
+const TransactionFailedField = styled(Field)`
+  flex-direction: column;
+  align-items: flex-start;
+  gap: 4px;
+`
+
+const TransactionLogMessage = styled(FieldValue)`
+  line-break: anywhere;
+  font-weight: 400;
+  font-size: 14px;
+  line-height: 15px;
+`
+
+const TransactionLogKey = styled(FieldKey)`
+  display: flex;
+  align-items: center;
+  gap: 7px;
+`
+
+export interface ITransactionDetailBase {
+  transactionTransformed: TransformedTransaction
   network: Network
   tokensByNetwork: Token[]
 }
 
-export const ExplorerTransactionDetail: FC<IExplorerTransactionDetail> = ({
+export interface ITransactionDetail extends ITransactionDetailBase {
+  transaction: Transaction
+  explorerTransaction?: never
+}
+
+export interface IExplorerTransactionDetail extends ITransactionDetailBase {
+  transaction?: never
+  explorerTransaction: IExplorerTransaction
+}
+
+export type TransactionDetailProps =
+  | ITransactionDetail
+  | IExplorerTransactionDetail
+
+export const TransactionDetail: FC<TransactionDetailProps> = ({
+  transaction,
   explorerTransaction,
-  explorerTransactionTransformed,
+  transactionTransformed,
   network,
   tokensByNetwork,
 }) => {
   const [bottomSheetOpen, setBottomSheetOpen] = useState(false)
-  const { action, date, displayName, actualFee, dapp } =
-    explorerTransactionTransformed
-  const isRejected = explorerTransaction.status === "REJECTED"
-  const isTransfer = isTokenTransferTransaction(explorerTransactionTransformed)
-  const isNFT = isNFTTransaction(explorerTransactionTransformed)
-  const isSwap = isSwapTransaction(explorerTransactionTransformed)
-  const isTokenMint = isTokenMintTransaction(explorerTransactionTransformed)
+  const { action, date, displayName, actualFee, dapp } = transactionTransformed
+  const isRejected =
+    explorerTransaction?.status === "REJECTED" ||
+    transaction?.status === "REJECTED"
+  const isTransfer = isTokenTransferTransaction(transactionTransformed)
+  const isNFT = isNFTTransaction(transactionTransformed)
+  const isNFTTransfer = isNFTTransferTransaction(transactionTransformed)
+  const isSwap = isSwapTransaction(transactionTransformed)
+  const isTokenMint = isTokenMintTransaction(transactionTransformed)
+  const theme = useTheme()
   const title = useMemo(() => {
     if (isTransfer || isTokenMint) {
-      const { amount, tokenAddress } = explorerTransactionTransformed
+      const { amount, tokenAddress } = transactionTransformed
       return (
         <TransferTitle
           action={action}
@@ -116,8 +172,8 @@ export const ExplorerTransactionDetail: FC<IExplorerTransactionDetail> = ({
           fallback={displayName}
         />
       )
-    } else if (isNFT) {
-      const { contractAddress, tokenId } = explorerTransactionTransformed
+    } else if (isNFT || isNFTTransfer) {
+      const { contractAddress, tokenId } = transactionTransformed
       /** ERC721 */
       return (
         <NFTTitle
@@ -130,17 +186,18 @@ export const ExplorerTransactionDetail: FC<IExplorerTransactionDetail> = ({
     }
     return displayName
   }, [
-    action,
-    displayName,
-    explorerTransactionTransformed,
-    isNFT,
-    isTokenMint,
     isTransfer,
+    isTokenMint,
+    isNFT,
+    isNFTTransfer,
+    displayName,
+    transactionTransformed,
+    action,
     network.id,
   ])
   const additionalFields = useMemo(() => {
-    if (isTransfer) {
-      const { fromAddress, toAddress } = explorerTransactionTransformed
+    if (isTransfer || isNFTTransfer) {
+      const { fromAddress, toAddress } = transactionTransformed
       return (
         <>
           <AccountAddressField
@@ -158,7 +215,7 @@ export const ExplorerTransactionDetail: FC<IExplorerTransactionDetail> = ({
     }
     if (isSwap) {
       const { fromTokenAddress, toTokenAddress, fromAmount, toAmount } =
-        explorerTransactionTransformed
+        transactionTransformed
       const negativeFromAmount = BigNumber.from(0).sub(fromAmount)
       return (
         <>
@@ -179,21 +236,28 @@ export const ExplorerTransactionDetail: FC<IExplorerTransactionDetail> = ({
     }
     return null
   }, [
-    explorerTransactionTransformed,
-    isSwap,
     isTransfer,
+    isNFTTransfer,
+    isSwap,
+    transactionTransformed,
     network.id,
     tokensByNetwork,
   ])
   const titleShowsTo =
-    isTransfer && (action === "SEND" || action === "TRANSFER")
-  const titleShowsFrom = isTransfer && action === "RECEIVE"
-  const displayContractAddress = formatTruncatedAddress(
-    explorerTransaction.contractAddress,
-  )
-  const displayTransactionHash = formatTruncatedAddress(
-    explorerTransaction.transactionHash,
-  )
+    (isTransfer || isNFTTransfer) &&
+    (action === "SEND" || action === "TRANSFER")
+  const titleShowsFrom = (isTransfer || isNFTTransfer) && action === "RECEIVE"
+  const displayContractAddress =
+    !!explorerTransaction &&
+    formatTruncatedAddress(explorerTransaction.contractAddress)
+  const hash = explorerTransaction?.transactionHash || transaction?.hash
+  const displayTransactionHash = !!hash && formatTruncatedAddress(hash)
+  const calls = explorerTransaction?.calls || transaction?.meta?.transactions
+  const errorMessage =
+    isRejected &&
+    transaction &&
+    getErrorMessageFromErrorDump(transaction.failureReason?.error_message)
+
   return (
     <StyledTransactionDetailWrapper
       title={
@@ -201,7 +265,7 @@ export const ExplorerTransactionDetail: FC<IExplorerTransactionDetail> = ({
           {!isNFT && (
             <MainTransactionIconContainer>
               <TransactionIcon
-                transaction={explorerTransactionTransformed}
+                transaction={transactionTransformed}
                 size={80}
                 outline
               />
@@ -217,8 +281,8 @@ export const ExplorerTransactionDetail: FC<IExplorerTransactionDetail> = ({
                 <PrettyAccountAddress
                   accountAddress={
                     titleShowsTo
-                      ? explorerTransactionTransformed.toAddress
-                      : explorerTransactionTransformed.fromAddress
+                      ? transactionTransformed.toAddress
+                      : transactionTransformed.fromAddress
                   }
                   networkId={network.id}
                   size={20}
@@ -230,34 +294,35 @@ export const ExplorerTransactionDetail: FC<IExplorerTransactionDetail> = ({
         </>
       }
     >
-      <TransactionCallDataBottomSheet
-        open={bottomSheetOpen}
-        onClose={() => setBottomSheetOpen(false)}
-        calls={explorerTransaction.calls}
-      />
+      {calls && (
+        <TransactionCallDataBottomSheet
+          open={bottomSheetOpen}
+          onClose={() => setBottomSheetOpen(false)}
+          calls={calls}
+        />
+      )}
       <ExpandableFieldGroup
         icon={
-          <TransactionIcon
-            transaction={explorerTransactionTransformed}
-            size={40}
-          />
+          <TransactionIcon transaction={transactionTransformed} size={40} />
         }
         title="Action"
         subtitle={displayName}
       >
-        <Field>
-          <FieldKey>Contract</FieldKey>
-          <FieldValue>
-            <StyledCopyIconButton
-              size="s"
-              copyValue={explorerTransaction.contractAddress}
-              variant="transparent"
-            >
-              {displayContractAddress}
-            </StyledCopyIconButton>
-          </FieldValue>
-        </Field>
-        {explorerTransaction.calls && (
+        {displayContractAddress && (
+          <Field>
+            <FieldKey>Contract</FieldKey>
+            <FieldValue>
+              <StyledCopyIconButton
+                size="s"
+                copyValue={explorerTransaction.contractAddress}
+                variant="transparent"
+              >
+                {displayContractAddress}
+              </StyledCopyIconButton>
+            </FieldValue>
+          </Field>
+        )}
+        {calls && (
           <Field clickable onClick={() => setBottomSheetOpen(true)}>
             <FieldKey>Call data</FieldKey>
             <FieldValue>
@@ -265,10 +330,10 @@ export const ExplorerTransactionDetail: FC<IExplorerTransactionDetail> = ({
             </FieldValue>
           </Field>
         )}
-        {!!explorerTransaction.events?.length && (
+        {!!explorerTransaction?.events?.length && (
           <SectionHeader>Event</SectionHeader>
         )}
-        {explorerTransaction.events.map((event, index) => {
+        {explorerTransaction?.events.map((event, index) => {
           const { name, address, parameters } = event
           const displayName = entryPointToHumanReadable(name)
           const displayAddress = formatTruncatedAddress(address)
@@ -312,23 +377,50 @@ export const ExplorerTransactionDetail: FC<IExplorerTransactionDetail> = ({
           <FieldKey>Status</FieldKey>
           <FieldValue>{isRejected ? "Failed" : "Complete"}</FieldValue>
         </Field>
+        {errorMessage && (
+          <Field>
+            <FieldKey>Reason</FieldKey>
+            <LeftPaddedField>{errorMessage}</LeftPaddedField>
+          </Field>
+        )}
         {dapp && <DappContractField knownContract={dapp} />}
         {additionalFields}
         {actualFee && <FeeField fee={actualFee} networkId={network.id} />}
       </FieldGroup>
-      <FieldGroup>
-        <Field
-          clickable
-          onClick={() =>
-            openVoyagerTransaction(explorerTransaction.transactionHash, network)
-          }
-        >
-          <FieldKey>Transaction ID</FieldKey>
-          <FieldValue>
-            <HyperlinkText>{displayTransactionHash}</HyperlinkText>
-          </FieldValue>
-        </Field>
-      </FieldGroup>
+      {isRejected && transaction && (
+        <FieldGroup>
+          <TransactionFailedField clickable>
+            <TransactionLogKey>
+              <div>Transaction log</div>
+              <CopyTooltip
+                message="Copied"
+                copyValue={
+                  transaction.failureReason?.error_message || hash || ""
+                }
+              >
+                <ContentCopyIcon style={{ fontSize: 12 }} />
+              </CopyTooltip>
+            </TransactionLogKey>
+            <TransactionLogMessage style={{ color: theme.text2 }}>
+              {transaction.failureReason?.error_message || "Unknown error"}
+            </TransactionLogMessage>
+          </TransactionFailedField>
+        </FieldGroup>
+      )}
+      {hash && (
+        <FieldGroup>
+          <Field
+            clickable
+            onClick={() => openVoyagerTransaction(hash, network)}
+          >
+            <FieldKey>Transaction ID</FieldKey>
+            <FieldValue>
+              <HyperlinkText>{displayTransactionHash}</HyperlinkText>
+            </FieldValue>
+          </Field>
+        </FieldGroup>
+      )}
+      <pre>{JSON.stringify({ transactionTransformed }, null, 2)}</pre>
     </StyledTransactionDetailWrapper>
   )
 }
