+++ packages/extension/src/ui/features/accountActivity/TransactionsListWrapper.tsx
@@ -1,0 +1,6 @@
+import styled from "styled-components"
+
+export const TransactionsListWrapper = styled.div`
+  display: flex;
+  flex-direction: column;
+`
